timestamp update

+++ terminal/heartbeat.go
@@ -3,9 +3,9 @@ package main
 import (
 	"k8s.io/client-go/kubernetes"
          metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-        "fmt"
         "time"
         "strconv"
+        "fmt"
 )
 
 //Heartbeater is
@@ -24,15 +24,23 @@ type terHeartbeater struct {
 
 func (t *terHeartbeater) CleanTerminalJob(clientSet *kubernetes.Clientset) {
     deploymentsClient := clientSet.AppsV1().Deployments(t.namespace)
-    deletePolicy := metav1.DeletePropagationForeground
-    if err := deploymentsClient.Delete(t.terminalID, &metav1.DeleteOptions{
-        PropagationPolicy: &deletePolicy,
-    }); err != nil {
-        panic(err)
+    result, err := deploymentsClient.Get(t.terminalID, metav1.GetOptions{})
+    if err != nil {
+       panic(err)
+    }
+    lastHeartbeat,_ := strconv.Atoi(result.Annotations["fist.seayun.com/lastHeartbeat"])
+    heartbeatTime := time.Now().Unix() - int64(lastHeartbeat)
+    if heartbeatTime > 600 { // time.Minute() * 10 = 600
+        deletePolicy := metav1.DeletePropagationForeground
+        if err := deploymentsClient.Delete(t.terminalID, &metav1.DeleteOptions{
+            PropagationPolicy: &deletePolicy,
+        }); err != nil {
+            panic(err)
+        }
     }
 }
 
-func (t *terHeartbeater) UpdateTimestamp(clientSet *kubernetes.Clientset) {
+func (t *terHeartbeater) UpdateTimestamp(clientSet *kubernetes.Clientset) error{
     deploymentsClient := clientSet.AppsV1().Deployments(t.namespace)
     result, err := deploymentsClient.Get(t.terminalID, metav1.GetOptions{})
     if err != nil {
@@ -40,5 +48,7 @@ func (t *terHeartbeater) UpdateTimestamp(clientSet *kubernetes.Clientset) {
     }
     timestamp := strconv.FormatInt( time.Now().Unix(), 10)
     result.Annotations["fist.seayun.com/lastHeartbeat"] = timestamp
-    fmt.Println(result.Annotations["fist.seayun.com/lastHeartbeat"])
+    _, updateErr := deploymentsClient.Update(result)
+    return updateErr
+
 }
